Git submodules are (ultimately) not worth using.
I've been a fan of the control and less object bloat, but they are
also annoying to maintain/update, and for lack of a better word, feel
like a "hack". Let's remove them!

This just deletes the file. The two libraries need to be added in
following commits.

!! BEWARE !!

You will most likely have to manually remove any submodule folders
for those next commits to apply cleanly.