There should be a way to emulate keyboard events in WebWorks BB10.
You can trigger the close/open events manually in the Events panel.

I also added an event to open and close the keyboard, which overlays
a crude keyboard placeholder onto the app, and triggers the
keyboardPosition event on each animation step. Putting all of this in
the Events panel felt the best for now, until the time comes when a
solid, worthwhile panel can be made for the virtual keyboard. I have put
this as a TODO note in the spec/events file as well.

To see the events in action, use the snippet below.

    <html><head><script>
        ["keyboardClosing",
         "keyboardClosed",
         "keyboardOpening",
         "keyboardOpened",
         "keyboardPosition",
        ].forEach(function (evt) {
            blackberry.event.addEventListener(evt, function () {
                    console.log(evt + " -> " + arguments[0]);
                });
        });
    </script></head></html>

GitHub Issue: https://github.com/blackberry/Ripple-UI/issues/387